refactor: /seminar api 폴더구조 수정
