+++ app/[locale]/community/seminar/[id]/page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from 'react';
 
-import { getSeminarPost } from '@/apis/seminar';
+import { getSeminarPost } from '@/apis/v1/seminar/[id]';
 import InvalidIDFallback from '@/components/common/InvalidIDFallback';
 import PostFallback from '@/components/layout/fallback/PostFallback';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
